20240105 SMH [메인] api 수정

+++ src/main/java/kr/co/maple/module/main/model/Top1UnionRankingDTO.java
@@ -11,6 +11,6 @@ public class Top1UnionRankingDTO {
 	private int characterLevel;
 	private String characterClass;
 	private int unionLevel;
-	private int unionPower;
+	private long unionPower;
 	private String characterImage;
 }
